Fix a clippy::unnecessary_cast warning

+++ src/scalar.rs
@@ -999,7 +999,7 @@ impl Scalar {
             6 => (256 + w - 1)/w as usize,
             7 => (256 + w - 1)/w as usize,
             // See comment in to_radix_2w on handling the terminal carry.
-            8 => (256 + w - 1)/w + 1 as usize,
+            8 => (256 + w - 1)/w + 1_usize,
             _ => panic!("invalid radix parameter"),
         };
